[to be squashed] Changed byte to int to avoid unneccecary casts

+++ ChatfuelTestTask/InputController.cs
@@ -11,11 +11,11 @@ namespace ChatfuelTestTask
 		public InputController( Lift lift ) => _lift = lift;
 
 		[HttpGet( "call-from/{floor}" )]
-		public Task CallFrom( byte floor )
+		public Task CallFrom( int floor )
 			=> _lift.CallFromOutside( floor );
 
 		[HttpGet( "direct-to/{floor}" )]
-		public Task DirectTo( byte floor )
+		public Task DirectTo( int floor )
 			=> _lift.DirectFromInside( floor );
     }
 }
